Fix typo in time passes
Summary:
Previously NamedRegionTimer's constructor was being called
with no local variable associated with it owing to a typo. We need a
local variable to keep track of the time spent in the scope. At the
end of the scope, the destructor will be called an then the timer will
stop.

(cherry picked from commit 8784dfbd4ea65c2e95accbb4232f0db2d20fd3fe)